Refactor adb client into library and binary.
This should simplify creating clients with different grpc options
